AF-1871: Improvements to SQL Datasource tests
