--- kie-soup-dataset/kie-soup-dataset-sql-tests/kie-soup-dataset-sql-tests-db2/pom.xml
@@ -1,55 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~ Copyright 2014 Red Hat, Inc. and/or its affiliates.
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~       http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <parent>
-    <groupId>org.kie.soup</groupId>
-    <artifactId>kie-soup-dataset-sql-tests</artifactId>
-    <version>7.19.0-SNAPSHOT</version>
-  </parent>
-
-  <artifactId>kie-soup-dataset-sql-tests-db2</artifactId>
-  <packaging>jar</packaging>
-
-  <name>KIE Soup Dataset DB2</name>
-
-  <properties>
-    <db2.driver.path>${basedir}/src/test/lib/db2jcc4.jar</db2.driver.path>
-  </properties>
-
-  <dependencyManagement>
-    <dependencies>
-      <dependency>
-        <groupId>com.ibm</groupId>
-        <artifactId>db2jcc4</artifactId>
-        <scope>system</scope>
-        <systemPath>${db2.driver.path}</systemPath>
-        <version>1</version>
-      </dependency>
-    </dependencies>
-  </dependencyManagement>
-
-  <dependencies>
-
-    <dependency>
-      <groupId>com.ibm</groupId>
-      <artifactId>db2jcc4</artifactId>
-    </dependency>
-  </dependencies>
-
-</project>
