--- kie-soup-dataset/kie-soup-dataset-sql-tests/kie-soup-dataset-sql-tests-db2/src/test/resources/testdb-db2.properties
@@ -1,5 +1,0 @@
-server=192.168.0.10
-database=TEST
-port=50000
-user=db2inst1
-password=password1!
