--- kie-soup-dataset/kie-soup-dataset-sql-tests/kie-soup-dataset-sql-tests-generic/pom.xml
@@ -1,136 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!-- ~ Copyright 2014 Red Hat, Inc. and/or its affiliates. ~ ~ Licensed under 
-  the Apache License, Version 2.0 (the "License"); ~ you may not use this file 
-  except in compliance with the License. ~ You may obtain a copy of the License 
-  at ~ ~ http://www.apache.org/licenses/LICENSE-2.0 ~ ~ Unless required by 
-  applicable law or agreed to in writing, software ~ distributed under the 
-  License is distributed on an "AS IS" BASIS, ~ WITHOUT WARRANTIES OR CONDITIONS 
-  OF ANY KIND, either express or implied. ~ See the License for the specific 
-  language governing permissions and ~ limitations under the License. -->
-
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <parent>
-    <groupId>org.kie.soup</groupId>
-    <artifactId>kie-soup-dataset-sql-tests</artifactId>
-    <version>7.20.0-SNAPSHOT</version>
-  </parent>
-
-  <artifactId>kie-soup-dataset-sql-tests-generic</artifactId>
-  <packaging>jar</packaging>
-
-  <name>KIE Soup Dataset Generic Test</name>
-
-  <properties>
-    <java.module.name>org.drools.persistence.jpa</java.module.name>
-
-    <!-- These are properties used in the database profiles. Some of them 
-      must be initialized to be empty so that Maven applies their values via filtering 
-      to the resources. -->
-    <maven.hibernate.dialect>org.hibernate.dialect.H2Dialect</maven.hibernate.dialect>
-    <maven.datasource.classname>org.h2.jdbcx.JdbcDataSource</maven.datasource.classname>
-    <maven.jdbc.driver.class>org.h2.Driver</maven.jdbc.driver.class>
-    <maven.jdbc.db.name />
-    <maven.jdbc.db.port />
-    <maven.jdbc.db.server />
-    <maven.jdbc.driver.jar />
-    <maven.jdbc.username />
-    <maven.jdbc.password />
-    <maven.jdbc.url>jdbc:h2:mem:test</maven.jdbc.url>
-    <maven.jdbc.schema>public</maven.jdbc.schema>
-    
-    <!-- versions for non managed dependencies - this should be changed-->
-    <version.narayana.jta>5.9.0.Final</version.narayana.jta>
-    <version.tomcat.dbcp>9.0.13</version.tomcat.dbcp>
-  </properties>
-
-  <dependencyManagement>
-    <dependencies>
-      <dependency>
-        <groupId>org.jboss.narayana.jta</groupId>
-        <artifactId>narayana-jta</artifactId>
-        <version>${version.narayana.jta}</version>
-      </dependency>
-      <dependency>
-        <groupId>org.apache.tomcat</groupId>
-        <artifactId>tomcat-dbcp</artifactId>
-        <version>${version.tomcat.dbcp}</version>
-      </dependency>
-    </dependencies>
-  </dependencyManagement>
-
-
-  <dependencies>
-
-    <dependency>
-      <groupId>com.h2database</groupId>
-      <artifactId>h2</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>junit</groupId>
-      <artifactId>junit</artifactId>
-    </dependency>
-
-    <dependency>
-      <groupId>commons-collections</groupId>
-      <artifactId>commons-collections</artifactId>
-    </dependency>
-
-    <dependency>
-      <groupId>javax.persistence</groupId>
-      <artifactId>javax.persistence-api</artifactId>
-      <optional>true</optional>
-    </dependency>
-    <dependency>
-      <groupId>org.jboss.spec.javax.transaction</groupId>
-      <artifactId>jboss-transaction-api_1.2_spec</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.jboss.narayana.jta</groupId>
-      <artifactId>narayana-jta</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.tomcat</groupId>
-      <artifactId>tomcat-dbcp</artifactId>
-    </dependency>
-    <!-- logging -->
-    <dependency>
-      <groupId>org.slf4j</groupId>
-      <artifactId>slf4j-api</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.jboss.logging</groupId>
-      <artifactId>jboss-logging</artifactId>
-    </dependency>
-
-  </dependencies>
-
-  <build>
-    <testResources>
-      <testResource>
-        <directory>src/test/resources</directory>
-        <filtering>false</filtering>
-      </testResource>
-      <testResource>
-        <directory>src/test/filtered-resources</directory>
-        <filtering>true</filtering>
-      </testResource>
-    </testResources>
-    <plugins>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-surefire-plugin</artifactId>
-        <configuration>
-          <additionalClasspathElements>
-            <additionalClasspathElement>${maven.jdbc.driver.jar}</additionalClasspathElement>
-          </additionalClasspathElements>
-        </configuration>
-      </plugin>
-
-    </plugins>
-  </build>
-
-</project>
