--- kie-soup-dataset/kie-soup-dataset-sql-tests/kie-soup-dataset-sql-tests-h2/src/test/resources/testdb-h2.properties
@@ -1,3 +1,0 @@
-url=jdbc:h2:~/test
-user=sa
-password=sa
