--- kie-soup-dataset/kie-soup-dataset-sql-tests/kie-soup-dataset-sql-tests-mariadb/src/test/resources/testdb-mariadb.properties
@@ -1,7 +1,0 @@
-server=db-28.rhev-ci-vms.eng.rdu2.redhat.com
-database=dballo19
-port=3306
-user=dballo19
-password=dballo19
-
-
