--- kie-soup-dataset/kie-soup-dataset-sql-tests/kie-soup-dataset-sql-tests-monetdb/src/test/resources/testdb-monetdb.properties
@@ -1,3 +1,0 @@
-url=jdbc:monetdb://104.155.41.56:50000/voc
-user=monetdb
-password=$monetdb123
