--- kie-soup-dataset/kie-soup-dataset-sql-tests/kie-soup-dataset-sql-tests-mysql/src/test/resources/testdb-mysql.properties
@@ -1,4 +1,0 @@
-url=jdbc:mysql://localhost:3306/dashbuilder
-user=root
-password=root
-
