--- kie-soup-dataset/kie-soup-dataset-sql-tests/kie-soup-dataset-sql-tests-oracle/src/test/resources/testdb-oracle.properties
@@ -1,3 +1,0 @@
-url=jdbc:oracle:thin:@dev151.mw.lab.eng.bos.redhat.com:1521:qaora12
-user=dballo11
-password=dballo11
