--- kie-soup-dataset/kie-soup-dataset-sql-tests/kie-soup-dataset-sql-tests-postgres/src/test/resources/testdb-postgres.properties
@@ -1,6 +1,0 @@
-server=db20.mw.lab.eng.bos.redhat.com
-database=dballo19
-port=5432
-user=dballo19
-password=dballo19
-
