--- kie-soup-dataset/kie-soup-dataset-sql-tests/kie-soup-dataset-sql-tests-sqlserver/src/test/java/org/dashbuilder/dataprovider/sql/SQLServerTest.java
@@ -1,31 +1,0 @@
-/*
- * Copyright 2015 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.dashbuilder.dataprovider.sql;
-
-import org.junit.Test;
-
-public class SQLServerTest extends SQLTestSuite {
-
-    @Override
-    protected DatabaseTestSettings createTestSettings() {
-        return new SQLServerTestSettings();
-    }
-
-    @Test
-    public void testAll() throws Exception {
-        super.testAll();
-    }
-}
