--- kie-soup-dataset/kie-soup-dataset-sql-tests/kie-soup-dataset-sql-tests-sqlserver/src/test/resources/testdb-sqlserver.properties
@@ -1,3 +1,0 @@
-url=jdbc:sqlserver://db18.mw.lab.eng.bos.redhat.com:1433;DatabaseName=dballo02
-user=dballo02
-password=dballo02
