--- kie-soup-dataset/kie-soup-dataset-sql-tests/kie-soup-dataset-sql-tests-sybase/src/test/resources/testdb-sybase.properties
@@ -1,5 +1,0 @@
-server=db05.mw.lab.eng.bos.redhat.com
-database=dballo01
-port=5000
-user=dballo01
-password=dballo01
