+++ kie-soup-dataset/kie-soup-dataset-sql-tests/src/test/filtered-resources/datasource.properties
